Move DatafeedClient out of datafeeds package
Before, the DatafeedClient is under symphony.api package, so moving it
to datafeeds package might break the compatibility of project. That's
why it should be moved out of datafeed package, back to symphony.api.
